django_sample/pyproject.toml
[tool.poetry]
name = "django-sample"
version = "0.1.0"
description = "Django Revolution sample project"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
django = "^5.2.4"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
pydantic = "^2.11.7"
rich = "^14.0.0"
pytest = "^8.4.1"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
django-cors-headers = "^4.3.1"
drf-nested-routers = "^0.93.4"

# Django Revolution
django-revolution = "*"
django-unfold = "^0.63.0"

[tool.poetry.group.dev.dependencies]
django-revolution = { path = "../django_revolution", develop = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
